% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{date_number}
\alias{date_number}
\title{Convert dates to day numbers}
\usage{
date_number(start_date, date_list)
}
\arguments{
\item{start_date}{object that can be coerced with as.Date}

\item{date_list}{vector that can be coerced with as.Date}
}
\value{
vector of integers
}
\description{
A vector of dates (or data that can be coerced as dates) are converted
to day number relative to start_date. E.g. if a date is equal to start_date
1 is returned.
}
\examples{
(dates <- seq.Date(Sys.Date(), by = 1, length.out = 10))
date_number(start_date = dates[1], date_list = dates)
}
