% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{format_Data2}
\alias{format_Data2}
\title{Create a matrix with all data for the Data2 matrix.}
\usage{
format_Data2(fish, env, ndays, missing_days = NULL)
}
\arguments{
\item{fish}{data frame with three mandatory columns "marked", "capture_day" and "recapture_day"}

\item{env}{data frame with two mandatory columns "w_temp" and "w_level".}

\item{ndays}{numeric number of days data have in operation, e.g. start date to stop date}

\item{missing_days}{numeric vector of day numbers when capture (and recapture) didn't work.}
}
\value{
a matrix with ndays rows and ndays + 6 columns
}
\description{
Return a matrix with data almost ready to be cut and pasted into the
main data file (Data2_1group.odc) for the model.
}
\details{
Each row data frame fish (with mandatory columns "marked", "capture_day" and "recapture_day") represent
a caught fish. Column capture_day contains the number of days between the start of the experiment
the day the fish was caught for the first time. Column recapture_day is the same if recaptured
and NA otherwise. Column marked is set to TRUE for all fish that was marked and subject
for for recapture and FALSE otherwise, e.g. when marked is false the fish was not marked
and not subject for possible recapture.

Data frame env must have ndays rows and two mandatory columns w_temp (water temperature)
and w_level (water level). No NAs allowed so impute the data if you have missing values.

Ndays is the number of days the experiment was running including any downtime.
Optional parameter missing_dats is a vector with day numbers when the experiment was put
on hiatus.
}
\examples{
ndays <- 5
(smolts <- data.frame(
  capture_day = c(2,2,3),
  recapture_day = c(NA, 3, 4),
  marked = c(FALSE, TRUE, TRUE)))
(envdata <- data.frame(
  w_temp = c(7, 7, 8, 8.5, 9),
  w_level = c(100, 110, 105, 150, 145)))

format_Data2(fish = smolts, env = envdata, ndays = ndays)

}
