% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{gen_Inits2}
\alias{gen_Inits2}
\title{Generate Inits2 data}
\usage{
gen_Inits2(n, llambda = 0.1, lphi = 1, g)
}
\arguments{
\item{n}{numeric number of values generated}

\item{llambda}{numeric value(s) that will be repeated to get n values (default .1)}

\item{lphi}{numeric value(s) that will be repeated to get n values (default 1)}

\item{g}{numeric value(s) that will be repeated to get n values}
}
\value{
numeric matrix n x 3 with column names 'llambda[,1]', 'lphi[,1]' and 'g[,1]'
}
\description{
Generate a matrix suitable to save into a Inits2cX_group1.odc file. This function
is used by save_bugsdata() and is probably not called directly by the user.
}
\examples{
gen_Inits2(3, g = g_inits_1)

}
