% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_exporter.R
\name{get_script}
\alias{get_script}
\title{Export a script}
\usage{
get_script(scriptname, outname = scriptname, overwrite = FALSE)
}
\arguments{
\item{scriptname}{character Name of (known) script}

\item{outname}{character Name of output (default = scriptname)}

\item{overwrite}{logical Overwrite existing file (default = FALSE)}
}
\value{
Invisible return the value of the file.copy() operation. As a side effect
a script file is saved
}
\description{
The package have a couple of useful(?) scripts that you can use as a starting
point to build scripts that reformat data to data suitable for the BlackBox
bugs-model. Use `get_script()` to copy a script to your own workspace.
}
\details{
Known scripts:
\itemize{
 \item{"sote_import2bugs.R"} {read and format a file in the Sötebasen import format
 (i.e. the format we submit to Sötebasen admin for import) to bugs}
 \item{"sote2bugs.R"} {read and format a file exported from Sötebasen (i.e. a file
 we get when we export data from Sötebasen) to bugs}
}
}
\examples{
\dontrun{
get_script("sote_import2bugs.R")
}
}
