% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{save_bugsdata}
\alias{save_bugsdata}
\title{Save an Excel file with Bugs data}
\usage{
save_bugsdata(x, path = ".", fname = "bugsdata.xlsx")
}
\arguments{
\item{x}{numeric matrix for Data2 file. See format_Data2()}

\item{path}{character path to create file in. Default current working directory}

\item{fname}{character file name of created file. Default "bugsdata.xlsx"}
}
\description{
This function will save an Excel file with all data needed to create all
files needed to run the smolt model in BlackBox. The file will have 6 sheets
named Data1, Data2, Inits1c1, Inits1c2, Inits2c1 and Inits2c2. Each sheet corresponds
to an .odc file in BlackBox and you must cut'n'paste from one format to the other.
}
\examples{
\dontrun{
save_bugsdata(Data2)
}

}
