% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{save_Rdatadump}
\alias{save_Rdatadump}
\title{Save bugsdata + some meta data as a R dump}
\usage{
save_Rdatadump(
  x,
  river,
  species,
  startd,
  stopd,
  missing_days = NULL,
  path = ".",
  fname = "RData_dump.RData"
)
}
\arguments{
\item{x}{numeric matrix for Data2 file. See format_Data2()}

\item{river}{character name of the river the data was collected}

\item{species}{character name of species}

\item{startd}{Date first date trap was running}

\item{stopd}{Date last date trap was running}

\item{missing_days}{numeric vector of day numbers when capture (and recapture) didn't work.}

\item{path}{character path to create file in. Default current working directory}

\item{fname}{character file name of created file. Default "RData_dump.RData"}
}
\description{
This function will save a R data dump with data needed for the smolt model + meta data
describing the mark/recapture experiment. This file can be loaded to analyze the data,
generate reports etc...
}
