% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetbase.R
\name{sdb}
\alias{sdb}
\title{Set or get values for options related to Sötebasen.}
\usage{
sdb(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Folder and filenames used to find data exported from Sötebasen on regular
basis. Default values points to locations on the Sölab restricted servers.
Change needed if you want to work on private copies of the data, for example
if you work remote without VPN.
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{root_folder} Datafiles root (//storage-dh.slu.se/restricted$/Sötebasen/Exporter)}
 \item{\code{occations} File with trapping occasions (SmoltfällaFångsttillfällen.csv)}
 \item{\code{catch_recatch} File with catch/recatch data (SmoltfällaÅterfångst.csv)}
 \item{\code{catch_envir}  File with temp/waterlevel data (SmoltfällaTemperatur.csv)}
}
}

