% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetbase.R
\name{sdb_read_catch_envir}
\alias{sdb_read_catch_envir}
\title{Read smolt trap environment data from Sötebasen export}
\usage{
sdb_read_catch_envir(VattenNamn = NULL, Year = NULL)
}
\arguments{
\item{VattenNamn}{return data only from river = VattenNamn, optional}

\item{Year}{return data only from Year, optional}
}
\value{
Return a data frame with columns:
\itemize{
 \item{InsamlingID}{}
 \item{AnsträngningID}{}
 \item{Datum}{}
 \item{WT}{}
 \item{WL}{}
 \item{Year}{}
 }
}
\description{
Data with all environment variables collected during trapping (`AnstrTyp` = "Smoltfälla") are read
from a predefined Sötebasen export file. The location of the file is predefined
but can be changed, see ?sdb()
}
\examples{
\dontrun{
sdb(root_folder = ".") # Use current folder instead of the predefined
catch <- sdb_read_catch_recatch()
}
}
