% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetbase.R
\name{sdb_read_catch_recatch}
\alias{sdb_read_catch_recatch}
\title{Read smolt catch/recatch data from Sötebasen export}
\usage{
sdb_read_catch_recatch(Art = NULL, VattenNamn = NULL, Year = NULL)
}
\arguments{
\item{Art}{return data only from Art = Art, optional}

\item{VattenNamn}{return data only from river = VattenNamn, optional}

\item{Year}{return data only from Year, optional}
}
\value{
Return a data frame with columns:
\itemize{
 \item{InsamlingID}{}
 \item{AnsträngningID}{}
 \item{FångstDatum}{}
 \item{FångstTid}{}
 \item{Art}{}
 \item{Längd1}{}
 \item{Vikt1}{}
 \item{Behandling}{}
 \item{MärkeNr}{}
 \item{StadiumKod}{}
 \item{AnmIndivid}{}
 \item{VattenNamn}{}
 \item{AnstrPlats}{}
 \item{Year}{}
 }
}
\description{
Data with all smolt catches defined with `AnstrTyp` = "Smoltfälla" are read
from a predefined Sötebasen export file. The location of the file is predefined
but can be changed, see ?sdb()
}
\examples{
\dontrun{
sdb(root_folder = ".") # Use current folder instead of the predefined
catch <- sdb_read_catch_recatch()
}
}
