% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smolt2bugs-package.R
\docType{package}
\name{smolts2bugs-package}
\alias{smolts2bugs}
\alias{smolts2bugs-package}
\title{smolts2bugs: Functions To Help Creation Of Datafiles For Blackbox Smoltmodel}
\description{
Functions To Help Creation Of Datafiles For Blackbox Smolt Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://slu-aqua-diadromous.github.io/smolt2bugs/}
}

}
\keyword{internal}
