% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_exporter.R
\name{get_script}
\alias{get_script}
\title{Export a script}
\usage{
get_script(scriptname, outname = scriptname, overwrite = FALSE)
}
\arguments{
\item{scriptname}{character Name of (known) script}

\item{outname}{character Name of output (default = scriptname)}

\item{overwrite}{logical Overwrite existing file (default = FALSE)}
}
\value{
Invisible return the value of the file.copy() operation.
}
\description{
The package have a couple of useful(?) scripts that you can use as a starting
point to build scripts that reformat data to data suitable for the BlackBox
bugs-model. Use `get_script()` to copy a script to your own workspace.
}
\details{
Known scripts:
\itemize{
 \item{"söte_import2bugs.R"} {read and format a file in the Sötebasen import format to bugs}
 \item{"söte2bugs.R"} {read and format a file exported from Sötebasen to bugs}
}
}
\examples{
if (FALSE) {
get_script("söte_import2bugs.R")
}
}
