% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetbase.R
\name{sdb}
\alias{sdb}
\title{Set or get limits for allowed values for some values}
\usage{
sdb(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get limits for allowed values for some values
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{root_folder}}{(\code{character};//storage-dh.slu.se/restricted$/Sötebasen/Exporter) Datafiles root }
 \item{\code{occations}}{(\code{character};SmoltfällaFångsttillfällen.csv) File with trapping occasions }
 \item{\code{catch_recatch}}{(\code{character};SmoltfällaÅterfångst.csv)  File with catch/recatch data}
 \item{\code{catch_envir}}{(\code{character};SmoltfällaTemperatur.csv) File with temp/waterlevel data}
}
}

