% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetbase.R
\name{sdb_read_occasions}
\alias{sdb_read_occasions}
\title{Read smolt trapping occasion data}
\usage{
sdb_read_occasions()
}
\value{
Return a data frame with columns:
\itemize{
 \item{InsamlingID}{}
 \item{AnsträngningID}{}
 \item{AnstrDatumStart}{}
 \item{AnstrDatumSlut}{}
 \item{AnstrTyp}{}
 \item{VattenNamn}{}
 \item{HaroID}{}
 \item{AnstrPlats}{}
 }
}
\description{
Data with all smolt trappings defined with `AnstrTyp` = "Smoltfälla" are read
form a predefined Sötebasen export file. The location of the file is predefined
but can be changed, see ?sdb()
}
\examples{
\dontrun{
sdb(root_folder = ".") # Use current folder instead of the predefined
occasions <- sdb_read_occasions()
}
}
